pyproject.toml
[project]
name = "shadow-apk-gateway"
version = "1.0.0"
description = "Universal APK-to-Gateway: Extract Android app API endpoints and expose them through a normalized, executable gateway"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "pydantic>=2.0,<3.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "cryptography>=42.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "frida>=16.0",
    "frida-tools>=12.0",
    "openapi-spec-validator>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "hypothesis>=6.0",
    "ruff>=0.3.0",
    "mypy>=1.8",
    "black>=24.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["packages*", "apps*"]

[tool.pytest.ini_options]
testpaths = ["packages", "apps", "tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM"]
ignore = ["S603", "S607"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.black]
line-length = 100
target-version = ["py311"]
